refactor: improve types, remove unused types; split
BREAKING CHANGE: Adds better types for inputs.  Removed unused Request and Insertion types.  If we want to add something like this back, we should add a DeliveryLog.

This PR also refactors index.ts into separate files.

TESTING=unit tests
